requirements.txt
cryptography==36.0.1
openpyxl==3.0.10
pandas==1.3.5
numpy==1.26.0
Pillow==9.5.0
pretty-html-table==0.9.16
psycopg2-binary
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
SQLAlchemy==1.4.39
uuid
virtualenv
WooCommerce
xlrd
XlsxWriter

